requirements.txt
# ============================================================================
# Dépendances - Gestion Couturier / An's Learning
# Compatible déploiement Render (Python 3.11+)
# ============================================================================
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
reportlab==4.0.7
qrcode[pil]==7.4.2
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
mysql-connector-python==8.3.0
openpyxl==3.1.2
bcrypt==4.1.2

